+++ lib/simple_gql.dart
@@ -1,6 +1,10 @@
 
 library simple_gql;
 
-export 'src/GQLResponse.dart';
-export 'src/GQLQuery.dart';
-export 'src/GQLMutation.dart' ;
+import 'dart:convert';
+import 'package:http/http.dart';
+import 'package:meta/meta.dart';
+
+part 'src/GQLClient.dart';
+part 'src/GQLResponse.dart';
+